db handler checks if name exists in before making changed

+++ db_handler.py
@@ -14,9 +14,17 @@ class DBHandler:
         #user = {"name":name,"IP":IP,"port":port }
 
         try:
-            user = {"_id":name,"IP":ip,"port":port }
-            self.user_collection.insert_one(user)
-            return True 
+            
+            if (self.user_collection.count_documents({"_id":name}) != 0):
+            
+                user = {"_id":name,"IP":ip,"port":port }
+                self.user_collection.insert_one(user)
+                return True 
+            
+            else:
+
+                print("USERNAME:" + str(name) + " DOES NOT EXIST!")
+                return False
 
         except:
 
@@ -27,9 +35,17 @@ class DBHandler:
     def remove_user(self,name):
 
         try:
-            user_query = {"_id":name}
-            self.user_collection.delete_one(user_query)
-            return True 
+
+            if (self.user_collection.count_documents({"_id":name}) != 0):
+                
+                user_query = {"_id":name}
+                self.user_collection.delete_one(user_query)
+                return True 
+            
+            else:
+
+                print("USERNAME:" + str(name) + " DOES NOT EXIST!")
+                return False
 
         except:
 
@@ -39,10 +55,18 @@ class DBHandler:
     def update_socket(self,name,ip,port):
 
         try:
-            user_query = {"_id":name}
-            user_newvalues = { "$set": {"IP": ip,"port": port}}
-            self.user_collection.update_one(user_query, user_newvalues)
-            return True 
+
+            if (self.user_collection.count_documents({"_id":name}) != 0):
+
+                user_query = {"_id":name}
+                user_newvalues = { "$set": {"IP": ip,"port": port}}
+                self.user_collection.update_one(user_query, user_newvalues)
+                return True 
+            
+            else:
+
+                print("USERNAME:" + str(name) + " DOES NOT EXIST!")
+                return False
         
         except:
             print("USERNAME:" + str(name) + "DOES NOT EXIST!")
@@ -53,11 +77,18 @@ class DBHandler:
 
         try: 
 
-            user_query = {"_id":name}
-            user_newvalues = {"$set": {"subjects": subjects}}
-            self.user_collection.update_one(user_query,user_newvalues)
+            if (self.user_collection.count_documents({"_id":name}) != 0):
+
+                user_query = {"_id":name}
+                user_newvalues = {"$set": {"subjects": subjects}}
+                self.user_collection.update_one(user_query,user_newvalues)
+                return True 
             
-            return True 
+            else:
+
+                print("USERNAME:" + str(name) + " DOES NOT EXIST!")
+                return False
+
         
         except:
             print("USERNAME:" + str(name) + "DOES NOT EXIST!")
@@ -98,28 +129,35 @@ class DBHandler:
     def retrieve_texts (self, name):
 
         try:
-            
-            user_cursor = self.user_collection.find_one({"_id":{"$eq":name}})
 
-            user_subjects = user_cursor["subjects"]
+            if (self.user_collection.count_documents({"_id":name}) != 0):
+            
+                user_cursor = self.user_collection.find_one({"_id":{"$eq":name}})
 
-            #for subject in user_subjects:
-            #    message = self.messages_collection.find_one({"subject":{"$eq":subject}})
-            #   print(message["text"])
+                user_subjects = user_cursor["subjects"]
+
+                #for subject in user_subjects:
+                #    message = self.messages_collection.find_one({"subject":{"$eq":subject}})
+                #   print(message["text"])
             
-            #print(user_subjects[0])
+                #print(user_subjects[0])
             
-            message_cursor = self.messages_collection.find({"subject":{"$in":user_subjects}})
-            #print(message_cursor["text"])
+                message_cursor = self.messages_collection.find({"subject":{"$in":user_subjects}})
+                #print(message_cursor["text"])
             
-            msg_list = []
+                msg_list = []
             
-            for document in message_cursor:
+                for document in message_cursor:
                 
-                msg = {"NAME":document["name"],"SUBJECT":document["subject"],"TEXT":document["text"]}
-                msg_list.append(msg)  
+                    msg = {"NAME":document["name"],"SUBJECT":document["subject"],"TEXT":document["text"]}
+                    msg_list.append(msg)  
             
-            return msg_list
+                return msg_list
+            
+            else:
+
+                print("USERNAME:" + str(name) + " DOES NOT EXIST!")
+                return 
         
         except:
